add IntProperty class and use properties for Recolor transition

+++ library/src/main/java/android/util/IntProperty.java
@@ -6,22 +6,22 @@ import android.os.Build;
 /**
  * Dummy class. Permits to extend same hidden class from android framework.
  * Actually in runtime will be used class from android framework and ObjectAnimator
- * optimizations for FloatProperty will be applied.
+ * optimizations for IntProperty will be applied.
  *
  * Created by Andrey Kulikov on 18.08.15.
  */
 @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH)
-public abstract class FloatProperty<T> extends Property<T, Float> {
+public abstract class IntProperty<T> extends Property<T, Integer> {
 
-    public FloatProperty(String name) {
-        super(Float.class, name);
+    public IntProperty(String name) {
+        super(Integer.class, name);
     }
 
-    public abstract void setValue(T object, float value);
+    public abstract void setValue(T object, int value);
 
     @Override
-    final public void set(T object, Float value) {
-        setValue(object, value);
+    final public void set(T object, Integer value) {
+        setValue(object, value.intValue());
     }
 
 }
